refactore movement scripts to parent from basic one
