+++ Assets/Scripts/PlayerMovement/BasicMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e438894d7d297f7409f06ad1695c983a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
